refactor(intelligence,http): move rag_ai_generator to mockforge-intelligence (#555 phase 10)
The RAG-based `AiGenerator` impl was one of the last AI-touching modules
still in `mockforge-http`. Both its foreign deps were originally
`mockforge-core` paths (`mockforge_core::ai_response::AiResponseConfig`,
`mockforge_core::openapi::response::AiGenerator`) but those have since
been promoted out of core:

- `AiResponseConfig` → `mockforge_foundation::ai_response`
- `AiGenerator` trait → `mockforge_openapi::response`
- `Result` / `Error::Config` → `mockforge_foundation`

That means intelligence can host this module without re-introducing the
core↔intelligence cycle that Issue #562 broke.

Changes:
- mockforge-intelligence/src/rag_ai_generator.rs (new) — verbatim move
  with imports rewritten to foundation + openapi paths; drops the now-
  unneeded `#![allow(deprecated)]` (the deprecation was on a function
  this file does not call)
- mockforge-http/src/rag_ai_generator.rs — now a 16-line re-export shim
  so `mockforge_http::rag_ai_generator::RagAiGenerator` callers keep
  resolving unchanged

Net: -549 LOC from mockforge-http, +566 LOC to mockforge-intelligence
(includes the 18 unit tests + 5 async tests that follow the module).

Verified:
- cargo build -p mockforge-intelligence -p mockforge-http
- cargo clippy -p mockforge-intelligence -p mockforge-http --all-targets -- -D warnings
- cargo test -p mockforge-intelligence (337/337 pass)
- cargo test -p mockforge-http (583/583 pass — was 601, -18 are now on the
  intelligence side)
- cargo fmt --all --check
